web: add stable releases since 2019-11-26 (#5)
* add Jitsi Meet releases after 2019-11-26 until 2020-04-10

* fix typo

* add Jitsi Videobridge releases

* add Jicofo releases

* drop links to Jitsi Meet changelogs

Individual CHANGELOGs are easily reachable via the release pages. Until someone is eager to do the additional work of copying the links out and inserting them here I suggest to drop the direct CHANGELOG links.

* move version numbers into link text

Since the 2020-04-01 release Jitsi Meet, Jicofo and the Videobridge are versioned individually. Keep only the release date in the release title and move individual release version next to the product name.

* fix videobridge URLs

* fix videobridge URLs

* fix URLs

* re-add build number

or whatever that number might represent

* add stable release 4468 from 2020-04-17

* add release 4546

* add release 4548